requirements.txt
google-genai == 1.28.0
chromadb == 1.0.15
load_dotenv
os
pandas == 2.3.1
python-docx == 1.2.0
fastapi == 0.116.1
uvicorn == 0.35.0
pydantic == 2.11.7